don't discard clientmaps while there's an open conn
If we discard a clientmap while there is a conn open from it, then
we put ourselves in a position to get out-of-sync on counting number
of concurrent connections in the DoS subsystem.

Now we continue to track clientmaps for these connections even when
there hasn't been a fresh connection lately; but we don't report
the non-fresh connections in our exported stats.

Fixes bug #40637.