demand the right number of descriptors during bootstrap
the old code waits to bootstrap until you have the first n+1 bridge
descriptors present, when you plan to use the first n bridges.

since we will soon change things to only fetch the descriptors for the
first n bridges, we want to be tighter in how many descriptors we demand
before proceeding to bootstrap.