when FetchBridgeDescsJIT, use more conservative guard set
When we're set to minimize the number of bridges we connect to, use
NumEntryGuards, not NumDirectoryGuards, even when choosing how many
bridges to choose between for directory fetches.

We prefer this more conservative number because we only fetch descriptors
for the first NumEntryGuards of them, so we want to avoid weird edge
conditions where we started with a cached version of a descriptor for
a later bridge, but we never update it, and it gets stale but we still
use it.