hs_cell: Add encryption overhead to calculation of padding target size
This fixes a bug that would result in us underestimating the size of our
encoded INTRODUCE1 messages, and adding more padding than we had intended.
This bug had been harmless until we fixed #40865, at which point it
caused us to make INTRODUCE1 messages that wouldn't fit into a
relay payload.